Move the two main traits to the top-level `hwtracer/lib.rs`.
Fundamentally, hwtracer exposes three traits: `Tracer`, `ThreadTracer`,
and `Trace`. They're small traits, and the things that any user of this
library needs to know about, so putting them in the library's top-level
file makes it easier for readers to see what they need to care about.
